Updating the login system for development purposes.
Previously, the login.js imported an array from fakeUsers.js and the
loging.js checked if the email and passwords are similar to one of the
registered users. The fakeUsers.js is substituted by fakeUsers.json and
the file is fetched in the head of each .js to show the data in the app.