Issue #19: Add IssueServiceTest.
The tests for the IssueService class is made in the IsseuServiceTest, in the test folder.

[Fix] Change the class UserRequest to have all attributes constructor. This brings advantages to handle data in the tests and service layers. This class is originally designed to exchange user information in the login/register.

[Possible issue] All the tests depends on the creation of user and project. Up to this point, projects and users are created with the userService and projectService classes. If one day any of these update, them, we have to update all the tests.

To test the Service layer, we don't need to know or care about how the persistence layer is implemented. Ideally, we should be able to write and test our Service layer code without wiring in our full persistence layer [1].

[1] https://www.baeldung.com/spring-boot-testing

[Missing] Test delete and update a project with one or more issues.